Move retry function to `utils`, make it nonprivate
