+++ analytics/analytics/job_processor/dimensions.py
@@ -1,14 +1,10 @@
-from dataclasses import dataclass
-import json
 from pathlib import Path
 import re
 from typing import Any
 
-from django.db import connections
 import gitlab
 import gitlab.exceptions
 from gitlab.v4.objects import ProjectJob
-import sentry_sdk
 import yaml
 
 from analytics.core.models.dimensions import (
@@ -21,6 +17,7 @@ from analytics.core.models.dimensions import (
     TimeDimension,
 )
 from analytics.job_processor.metadata import JobMiscInfo, NodeInfo, PackageInfo, PodInfo
+from analytics.job_processor.utils import get_job_retry_data
 
 UNNECESSARY_JOB_REGEX = re.compile(r"No need to rebuild [^,]+, found hash match")
 BUILD_STAGE_REGEX = r"^stage-\d+$"
@@ -30,80 +27,6 @@ class UnrecognizedJobType(Exception):
     def __init__(self, job_id: int, name: str) -> None:
         message = f"Unrecognized job type for Job: ({job_id}) {name}"
         super().__init__(message)
-
-
-@dataclass(frozen=True)
-class RetryInfo:
-    is_retry: bool
-    is_manual_retry: bool
-    attempt_number: int
-    final_attempt: bool
-
-
-def _job_retry_data(
-    job_id: int, job_name: str, job_commit_id: int, job_failure_reason: str
-) -> RetryInfo:
-    with connections["gitlab"].cursor() as cursor:
-        # the prior attempts for a given job are all jobs with a lower id, the same commit_id, and
-        # the same name. the commit_id is the foreign key for the pipeline.
-        # it's important to filter for a lower job id in the event that the webhook is delayed or
-        # received out of order.
-        cursor.execute(
-            """
-            SELECT COUNT(*)
-            FROM ci_builds
-            WHERE id < %(job_id)s
-            AND commit_id = %(commit_id)s
-            AND name = %(job_name)s
-            """,
-            {"job_id": job_id, "job_name": job_name, "commit_id": job_commit_id},
-        )
-        attempt_number = cursor.fetchone()[0] + 1
-
-        cursor.execute(
-            """
-            SELECT bm.config_options->>'retry'
-            FROM ci_builds_metadata bm
-            WHERE bm.build_id = %(job_id)s
-            """,
-            {"job_id": job_id},
-        )
-        job = cursor.fetchone()
-        if not job[0]:
-            # config_options->>retry should always be defined for non-trigger (aka Ci::Bridge)
-            # jobs in spack. this is an edge case where a job in gitlab isn't explicitly
-            # configured for retries at all.
-            sentry_sdk.capture_message(f"Job {job_id} missing retry configuration.")
-            # this is the default retry configuration for gitlab
-            # see https://docs.gitlab.com/ee/ci/yaml/#retry
-            retry_config = {
-                "max": 0,
-                "when": ["always"],
-            }
-        else:
-            retry_config = json.loads(job[0])
-
-        retry_max = retry_config["max"]
-        # A non-retryable job can either have an explicit max of zero, or no max at all.
-        # If the job is not retryable, the 'when' key will not exist
-        if retry_max in (0, None):
-            retry_reasons = []
-        # If the job is retryable, the 'when' key will be a list of reasons to retry
-        else:
-            retry_reasons = retry_config["when"]
-        # final_attempt is defined as an attempt that won't be retried for the retry_reasons
-        # or because it's gone beyond the max number of retries.
-        retryable_by_reason = "always" in retry_reasons or job_failure_reason in retry_reasons
-        retryable_by_number = attempt_number <= retry_max
-        final_attempt = not (retryable_by_reason and retryable_by_number)
-
-        return RetryInfo(
-            is_retry=attempt_number > 1,
-            # manual retries are all retries that are not part of the original job
-            is_manual_retry=attempt_number > retry_max + 1,
-            attempt_number=attempt_number,
-            final_attempt=final_attempt,
-        )
 
 
 def _assign_error_taxonomy(job_input_data: dict[str, Any], job_trace: str):
@@ -200,7 +123,7 @@ def create_job_data_dimension(
     job_name = job_input_data["build_name"]
     job_commit_id = job_input_data["commit"]["id"]
     job_failure_reason = job_input_data["build_failure_reason"]
-    retry_info = _job_retry_data(
+    retry_info = get_job_retry_data(
         job_id=job_id,
         job_name=job_name,
         job_commit_id=job_commit_id,
